Use xxhash for dstr_hash().
make dstr_hash and callers return size_t which is
guaranteed to be same size as platform (32 or 64)
Some dstring.hpp tuning

+++ include/dstr/dstr.h
@@ -205,7 +205,7 @@ void dstr_lstrip_sz(DSTR p, const char* selectors);
 void dstr_rstrip_sz(DSTR p, const char* selectors);
 
 /* hash for use in hash table */
-unsigned long dstr_hash(CDSTR src, int seed /*=0*/);
+size_t dstr_hash(CDSTR src, int seed /*=0*/);
 
 /*   atoi integer conversion
  *
